requirements.txt
certifi==2025.6.15
charset-normalizer==3.4.2
idna==3.10
mutagen==1.47.0
python-dotenv==1.1.1
requests==2.32.4
yt-dlp==2025.6.25

-e .